# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Service Ticket Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{id}/permissions:
    patch:
      operationId: updateUserPermissions
      tags: [users]
      summary: Update which dashboard sections a user can access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionsUpdate"
      responses:
        "200":
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List all data uploads
      responses:
        "200":
          description: List of uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Upload"
    post:
      operationId: uploadData
      tags: [uploads]
      summary: Upload Excel data file (active tickets, closed tickets, or MRF)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: Upload processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{id}:
    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an upload and its associated data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List all configured report schedules
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a new report delivery schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get a schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    User:
      type: object
      required: [id, name, email, role, isActive, permissions]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, ash, rsh, service_partner]
        isActive:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        managerId:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, manager, ash, rsh, service_partner]
        managerId:
          type: ["integer", "null"]
        permissions:
          type: array
          items:
            type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, ash, rsh, service_partner]
        isActive:
          type: boolean
        managerId:
          type: ["integer", "null"]
        permissions:
          type: array
          items:
            type: string
    PermissionsUpdate:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            type: string
    Upload:
      type: object
      required: [id, filename, fileType, recordCount, uploadedAt, status]
      properties:
        id:
          type: integer
        filename:
          type: string
        fileType:
          type: string
          enum: [active_tickets, closed_tickets, mrf_data]
        recordCount:
          type: integer
        uploadedAt:
          type: string
        status:
          type: string
          enum: [processing, completed, failed]
        errorMessage:
          type: ["string", "null"]
    UploadInput:
      type: object
      required: [fileType]
      properties:
        fileType:
          type: string
          enum: [active_tickets, closed_tickets, mrf_data]
        file:
          type: string
          format: binary
...
    Schedule:
      type: object
      required: [id, name, reportTypes, frequency, audiences, productCategories, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        reportTypes:
          type: array
          items:
            type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, custom]
        weekDay:
          type: ["integer", "null"]
        monthDay:
          type: ["integer", "null"]
        customCron:
          type: ["string", "null"]
        audiences:
          type: array
          items:
            type: string
            enum: [service_partner, ash, rsh, management]
        productCategories:
          type: array
          items:
            type: string
        filters:
          type: object
          properties:
            dateRangeDays:
              type: ["integer", "null"]
            servicePartner:
              type: ["string", "null"]
            ash:
              type: ["string", "null"]
            rsh:
              type: ["string", "null"]
        isActive:
          type: boolean
        lastRunAt:
          type: ["string", "null"]
        nextRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ScheduleInput:
      type: object
      required: [name, reportTypes, frequency, audiences, productCategories]
      properties:
        name:
          type: string
        reportTypes:
          type: array
          items:
            type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, custom]
        weekDay:
          type: ["integer", "null"]
        monthDay:
          type: ["integer", "null"]
        customCron:
          type: ["string", "null"]
        audiences:
          type: array
          items:
            type: string
            enum: [service_partner, ash, rsh, management]
        productCategories:
          type: array
          items:
            type: string
        filters:
          type: object
          properties:
            dateRangeDays:
              type: ["integer", "null"]
            servicePartner:
              type: ["string", "null"]
            ash:
              type: ["string", "null"]
            rsh:
              type: ["string", "null"]
        isActive:
          type: boolean
    ScheduleUpdate:
      type: object
      properties:
        name:
          type: string
        reportTypes:
          type: array
          items:
            type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, custom]
        weekDay:
          type: ["integer", "null"]
        monthDay:
          type: ["integer", "null"]
        customCron:
          type: ["string", "null"]
        audiences:
          type: array
          items:
            type: string
            enum: [service_partner, ash, rsh, management]
        productCategories:
          type: array
          items:
            type: string
        filters:
          type: object
          properties:
            dateRangeDays:
              type: ["integer", "null"]
            servicePartner:
              type: ["string", "null"]
            ash:
              type: ["string", "null"]
            rsh:
              type: ["string", "null"]
        isActive:
          type: boolean
...