# lib/api-zod/src/generated/types/scheduleUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Service Ticket Reporting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleUpdateAudiencesItem } from './scheduleUpdateAudiencesItem';
import type { ScheduleUpdateFilters } from './scheduleUpdateFilters';
import type { ScheduleUpdateFrequency } from './scheduleUpdateFrequency';

export interface ScheduleUpdate {
  name?: string;
  reportTypes?: string[];
  frequency?: ScheduleUpdateFrequency;
  /** @nullable */
  weekDay?: number | null;
  /** @nullable */
  monthDay?: number | null;
  /** @nullable */
  customCron?: string | null;
  audiences?: ScheduleUpdateAudiencesItem[];
  productCategories?: string[];
  filters?: ScheduleUpdateFilters;
  isActive?: boolean;
}